# [3.0] Perf: Skip visiting table inside ColumnExpression