… test/EFCore.SqlServer.FunctionalTests/Query/GearsOfWarQuerySqlServerTest.cs
… test/EFCore.SqlServer.FunctionalTests/Query/OwnedQuerySqlServerTest.cs
… test/EFCore.SqlServer.FunctionalTests/Query/QueryNavigationsSqlServerTest.cs
… test/EFCore.SqlServer.FunctionalTests/Query/SimpleQuerySqlServerTest.ResultOperators.cs
… test/EFCore.SqlServer.FunctionalTests/Query/SimpleQuerySqlServerTest.Select.cs
… test/EFCore.SqlServer.FunctionalTests/Query/SimpleQuerySqlServerTest.cs